Added Windows Platform 	* Created an entry and main windows function. It does most of 	  the tests that linux does (Threads is omitted since Windows 	  API contains functions, unlike Linux in where pthreads is off 	  the tables.) 	* Created a batch file to build via MSVC or Zig (zig cc). 		* Unfortunately, Zig 0.13.0 is the current support, as 		  Zig 0.14.0 became a massive headache (May be impossible 		  as of now. I'll wait to see 0.14.1 or 0.15.0 to retry) 		* This also contains setup_cl.bat from 4coder's repository. 		  Therefore, this will retain its MIT License, unlike the 		  rest of the repository which is UNLICENSE. 	* Added two defines for both creating (safely) and assigning DLL 	  functions 	* Added a maximum for s8_chr and s8_rchr so it doesn't go past 	  the stated buffer s8 length 	* Made some function declarations more consistent with others. 	* Other miscellaneous changes to ensure compilations for three 	  compilers (Linux Clang, Windows MSVC and Cross-Compiling Zig cc

+++ include/mem/memfn.h
@@ -12,7 +12,9 @@
 //                      appropriate on a project basis (Jobs via Threading)
 
 local inline usz
-rotr(usz x, int k) {
+rotr(usz x,
+     int k)
+{
 	// NOTE [32-Bit]: Bad for 32-bit support
 	return (x >> k) | (x << (64 - k));
 }
@@ -566,7 +568,7 @@ memsetu_usz(void *dst,
 	size -= pad;
 
 	while (pad--) {
-		d.pu8[pad] = val;
+		d.pu8[pad] = (u8)val;
 		val = rotr(val, 8);
 	}
